Lesson 5

+++ CardStorageService/CardStorageService.API/Controllers/CardsController.cs
@@ -1,7 +1,9 @@
-﻿using CardStorageService.API.Models.Requests;
+﻿using AutoMapper;
+using CardStorageService.API.Models.Requests;
 using CardStorageService.API.Models.Responses;
 using CardStorageService.Core.Interfaces;
 using CardStorageService.Storage.Models;
+using FluentValidation;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -14,11 +16,31 @@ namespace CardStorageService.API.Controllers
     {
         private readonly ILogger<CardsController> _logger;
         private readonly ICardService _service;
+        private readonly IMapper _mapper;
 
-        public CardsController(ILogger<CardsController> logger, ICardService service)
+        private readonly IValidator<CardCreateRequest> _cardCreateRequestValidator;
+        private readonly IValidator<CardDeleteRequest> _cardDeleteRequestValidator;
+        private readonly IValidator<CardGetByClientIdRequest> _cardGetByClientIdRequestValidator;
+        private readonly IValidator<CardGetByIdRequest> _cardGetByIdRequestValidator;
+        private readonly IValidator<CardUpdateRequest> _cardUpdateRequestValidator;
+
+        public CardsController(
+            ILogger<CardsController> logger,
+            ICardService service, IMapper mapper,
+            IValidator<CardCreateRequest> cardCreateRequestValidator,
+            IValidator<CardDeleteRequest> cardDeleteRequestValidator,
+            IValidator<CardGetByClientIdRequest> cardGetByClientIdRequestValidator,
+            IValidator<CardGetByIdRequest> cardGetByIdRequestValidator,
+            IValidator<CardUpdateRequest> cardUpdateRequestValidator)
         {
             _logger = logger;
             _service = service;
+            _mapper = mapper;
+            _cardCreateRequestValidator = cardCreateRequestValidator;
+            _cardDeleteRequestValidator = cardDeleteRequestValidator;
+            _cardGetByClientIdRequestValidator = cardGetByClientIdRequestValidator;
+            _cardGetByIdRequestValidator = cardGetByIdRequestValidator;
+            _cardUpdateRequestValidator = cardUpdateRequestValidator;
         }
 
         [HttpPost("create")]
@@ -27,14 +49,14 @@ namespace CardStorageService.API.Controllers
         {
             try
             {
-                var cardId = await _service.Create(new()
+                var validationResult = _cardCreateRequestValidator.Validate(request);
+                if (!validationResult.IsValid)
                 {
-                    ClientId = request.ClientId,
-                    CardNo = request.CardNo,
-                    ExpDate = request.ExpDate,
-                    Name = request.Name,
-                    CVV2 = request.CVV2
-                }, cts);
+                    return BadRequest(validationResult.ToDictionary());
+                }
+
+                var cardId = await _service.Create(_mapper.Map<Card>(request), cts);
+
                 return Ok(new CardCreateResponse()
                 {
                     CardId = cardId.ToString()
@@ -57,6 +79,12 @@ namespace CardStorageService.API.Controllers
         {
             try
             {
+                var validationResult = _cardDeleteRequestValidator.Validate(request);
+                if (!validationResult.IsValid)
+                {
+                    return BadRequest(validationResult.ToDictionary());
+                }
+
                 var count = await _service.Delete(request.CardId, cts);
                 return Ok(new CardDeleteResponse()
                 {
@@ -103,6 +131,12 @@ namespace CardStorageService.API.Controllers
         {
             try
             {
+                var validationResult = _cardGetByClientIdRequestValidator.Validate(request);
+                if (!validationResult.IsValid)
+                {
+                    return BadRequest(validationResult.ToDictionary());
+                }
+
                 var cards = await _service.GetByClientId(request.ClientId, cts);
                 return Ok(new CardGetByClientIdResponse()
                 {
@@ -126,6 +160,12 @@ namespace CardStorageService.API.Controllers
         {
             try
             {
+                var validationResult = _cardGetByIdRequestValidator.Validate(request);
+                if (!validationResult.IsValid)
+                {
+                    return BadRequest(validationResult.ToDictionary());
+                }
+
                 var card = await _service.GetById(request.CardId, cts);
                 return Ok(new CardGetByIdResponse()
                 {
@@ -149,15 +189,13 @@ namespace CardStorageService.API.Controllers
         {
             try
             {
-                var count = await _service.Update(new Card()
+                var validationResult = _cardUpdateRequestValidator.Validate(request);
+                if (!validationResult.IsValid)
                 {
-                    CardId = Guid.Parse(request.CardId),
-                    CardNo = request.CardNo,
-                    ClientId = request.ClientId,
-                    CVV2 = request.CVV2,
-                    ExpDate = request.ExpDate,
-                    Name = request.Name
-                }, cts);
+                    return BadRequest(validationResult.ToDictionary());
+                }
+
+                var count = await _service.Update(_mapper.Map<Card>(request), cts);
 
                 return Ok(new CardUpdateResponse()
                 {
